Split `resolve` module into submodules
