+++ packages/graph/hash_graph/lib/graph/src/store/postgres/resolve/context.rs
@@ -1,0 +1,59 @@
+use async_trait::async_trait;
+use error_stack::{Context, Result, ResultExt};
+use type_system::uri::VersionedUri;
+
+use crate::store::{
+    postgres::{
+        ontology::OntologyDatabaseType,
+        resolve::{ontology, ontology::RecordStream, Record},
+    },
+    AsClient, PostgresStore, QueryError,
+};
+
+/// Context used for [`Resolve`].
+///
+/// This is only used as an implementation detail inside of the [`postgres`] module.
+///
+/// [`Resolve`]: crate::store::query::Resolve
+/// [`postgres`]: super
+// TODO: Use the context to hold query data
+//   see https://app.asana.com/0/0/1202884883200946/f
+#[async_trait]
+pub trait PostgresContext {
+    async fn read_all_ontology_types<T>(&self) -> Result<RecordStream<T>, QueryError>
+    where
+        T: OntologyDatabaseType + TryFrom<serde_json::Value, Error: Context>;
+
+    async fn read_versioned_ontology_type<T>(
+        &self,
+        uri: &VersionedUri,
+    ) -> Result<Record<T>, QueryError>
+    where
+        T: OntologyDatabaseType + TryFrom<serde_json::Value, Error: Context>;
+}
+
+#[async_trait]
+impl<C: AsClient> PostgresContext for PostgresStore<C> {
+    async fn read_all_ontology_types<T>(&self) -> Result<RecordStream<T>, QueryError>
+    where
+        T: OntologyDatabaseType + TryFrom<serde_json::Value, Error: Context>,
+    {
+        Ok(ontology::row_stream_to_record_stream(
+            ontology::read_all_types(&self.client, T::table())
+                .await
+                .attach_printable("could not read ontology types")?,
+        ))
+    }
+
+    async fn read_versioned_ontology_type<T>(
+        &self,
+        uri: &VersionedUri,
+    ) -> Result<Record<T>, QueryError>
+    where
+        T: OntologyDatabaseType + TryFrom<serde_json::Value, Error: Context>,
+    {
+        ontology::read_versioned_type(&self.client, T::table(), uri)
+            .await
+            .attach_printable("could not read ontology type")
+    }
+}
