+++ packages/graph/hash_graph/lib/graph/src/store/postgres/resolve/entity.rs
@@ -1,0 +1,2 @@
+// TODO: Implement structural queries for entities
+//   see https://app.asana.com/0/0/1202510174412977/f
