+++ packages/graph/hash_graph/lib/graph/src/store/postgres/resolve/links.rs
@@ -1,0 +1,2 @@
+// TODO: Implement structural queries for links
+//   see https://app.asana.com/0/0/1202912966917503/f
